Fix problems with GitPython 2.0
Up to now, we specified GitPython >= 1.0.2 which included v2.0 which
changed its implementation details so it broke our error handling.
Now we don't depend on GitPython's `GitCommandError` any more as
there was no real need to subclass it anyway.

Not fully tested on my machine, will see if this actually fixes it.

See #56