Remove scripts from setup.py
There is already a console_scripts entrypoint which results in a script.
When installing to the system this results in /usr/bin/gitup.py and
/usr/bin/git-up.

Previously it was needed but 222cc6a9910f4fc44fd15a64da5db52a94d9a3c3
removed this requirement by using the full path.