Fix running PyGitUp from a subdirectory
Should fix #59. The bugfix to #58 was faulty when PyGitUp was run from
a subfolder as `git rev-parse --git-common-dir` is not equivalent to
`git rev-parse --show-toplevel` when run in a subfolder as it returns
a relative path in cases like this. Now `--git-common-dir` is used when
executed in a worktree, otherwise `--show-toplevel` is used.