refactor: drop the click dependency (#96)
This reduces the number of external dependencies and relies more on
Python's standard library.

It drops the --no-push option since argparse doesn't support having a
--option / --no-option. That's also the default and this makes it
clearer to the user what the default is.

The help text's formatting is slightly different. Mostly visible are
that the description and epilog are no longer indented.